Revert "Hotfix/banner"
